fix(model): Fix model attributes

+++ server/src/schema/pant.ts
@@ -4,19 +4,16 @@ export const typeDefs = gql`
   extend type Query {
     allPants: [Pant]
   }
+
   extend type Mutation {
-    createPant(
-      name: String!
-      color: String!
-      width: Float!
-      length: Float!
-    ): Pant
+    createPant(name: String!, color: String!, width: Int!, length: Int!): Pant
   }
+
   type Pant {
     id: ID!
     name: String!
     color: String!
-    width: Float!
-    length: Float!
+    width: Int!
+    length: Int!
   }
 `;
